pyproject.toml
[project]
name = "shadow-bureau-backend"
version = "1.0.0"
description = "Noir-themed campus intelligence platform"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "backboard-sdk>=1.5.0",
    "fastapi>=0.129.0",
    "google-generativeai>=0.8.6",
    "httpx>=0.28.1",
    "imagehash>=4.3.2",
    "neo4j>=6.1.0",
    "numpy>=2.4.2",
    "opencv-python-headless>=4.13.0.92",
    "pillow>=12.1.1",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "uvicorn[standard]>=0.40.0",
    "websockets>=16.0",
]
